Split out PartitionsBuilderSplitsFirstTest
